Add features to analyze HTTP packets
With this commit, a feature has been added that captures raw packets
from a network interface or a pcap file and extracts HTTP packets.
After the HTTP packets are parsed, the User-Agent information is
extracted and bandwidth measurement is performed for each User-Agent.
Additionally, a print method has been added to optionally print out
extracted information.